requirements.txt
numpy==1.21.0
opencv-python==4.5.3.56
PyQt5==5.15.4
scikit-image==0.18.2
tensorflow==2.5.0
Pillow==8.3.1
rembg==2.0.30
matplotlib==3.4.2
scipy==1.7.0
